Fix a bug where Par is considered a member of itself
Due to a breaking change in the sanctuary-type-identifiers spec[1],
Fluture adheres to both specs during a transition period.

The only flaw with that approach is that the new
sanctuary-type-identifiers considers constructors to be members of
themselves, and as such, an extra check is needed when identifying
Fluture-exported types using sanctuary-type-identifiers 3.

This extra check can be removed once Fluture drops support for
sanctuary-type-identifiers 2.

[1]: https://github.com/sanctuary-js/sanctuary-type-identifiers/releases/tag/v3.0.0